feat: publish to websocket subs afer height sync
